fix: revert changes to typehints in public API
Since `BcryptTrait` is a public API, changes to its method signatures can have a BC impact.
In particular, the change to `checkPassword` for 1.7.0 was incorrect, as `$user` is an array, and because the signature is intended to match those in the oauth2-server sources.

Additionally, some of the parameter annotations in oauth2-server class extensions were incorrect, and have been updated.

Fixes #36

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>